fix(storage): resolve slot collision between governanceMultisig and s_l1FeeCoefficient
- Removed all unused `pending` variables.
- Previously, `governanceMultisig` in CommitReveal2 and `s_l1FeeCoefficient` in CommitReveal2L2
  overlapped in the same storage slot, causing all tests to fail.
- Changed `s_l1FeeCoefficient` type from `uint8` to `uint256` to align with EVM slot boundaries
  and eliminate storage collision.
- After this fix, all test cases now pass successfully.

+++ src/governance/ICommitReveal2Governance.sol
@@ -34,5 +34,5 @@ interface ICommitReveal2Governance {
 }
 
 interface ICommitReveal2L2Governance {
-    function setL1FeeCoefficient(uint8 coefficient) external;
+    function setL1FeeCoefficient(uint256 coefficient) external;
 }
